refactor: 상태 관리 구조 개선 - BaseModel 도입, 환경변수 분리, 데이터/처리 로직 분리
